tools/sharepoint-sync/requirements.txt
python-frontmatter>=1.1.0
python-docx>=1.1.0
markdown>=3.5.0
pyyaml>=6.0.1
